+++ Team6/Team6/Controllers/CriminalController.cs
@@ -8,22 +8,32 @@ using System.Web;
 using System.Web.Mvc;
 using Team6.Models;
 using Team6.DAL;
+using System.IO;
 
 namespace Team6.Controllers
 {
     public class CriminalController : Controller
     {
         private Team6Context db = new Team6Context();
+        //FileStream fs = new FileStream("C:/Users/Muhammad Naviwala/Documents/GitHub/criminal_queries.log", FileMode.OpenOrCreate);
 
         // GET: /Criminal/
         public ActionResult Index()
         {
-            return View(db.Criminals.ToList());
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
+            return View(db.Criminals.OrderBy(x => x.FirstName).ThenBy(x => x.LastName).ToList());
         }
 
         // GET: /Criminal/Details/5
         public ActionResult Details(int? id)
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +49,10 @@ namespace Team6.Controllers
         // GET: /Criminal/Create
         public ActionResult Create()
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             return View();
         }
 
@@ -49,6 +63,10 @@ namespace Team6.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include="CriminalID,FirstName,LastName,EyeColor,Weight,Height,Gender,Ssn,Alias,HairColor,KnownAffiliates,DateOfBirth,Race,Address,State,ZipCode,PhoneNumber,misc")] Criminal criminal)
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Criminals.Add(criminal);
@@ -62,6 +80,10 @@ namespace Team6.Controllers
         // GET: /Criminal/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -81,6 +103,10 @@ namespace Team6.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include="CriminalID,FirstName,LastName,EyeColor,Weight,Height,Gender,Ssn,Alias,HairColor,KnownAffiliates,DateOfBirth,Race,Address,State,ZipCode,PhoneNumber,misc")] Criminal criminal)
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(criminal).State = EntityState.Modified;
@@ -93,6 +119,10 @@ namespace Team6.Controllers
         // GET: /Criminal/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -110,6 +140,10 @@ namespace Team6.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
             Criminal criminal = db.Criminals.Find(id);
             db.Criminals.Remove(criminal);
             db.SaveChanges();
@@ -123,6 +157,134 @@ namespace Team6.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        public ActionResult SearchCriminal(string firstName, string lastName, string weight, string height, string ssn, string address, string zipCode, Race race, State state, Gender gender, EyeColor eyeColor, HairColor hairColor)
+        {
+            if (!(System.Web.HttpContext.Current.User != null && System.Web.HttpContext.Current.User.Identity.IsAuthenticated))
+            {
+                return RedirectToAction("LogOn", "Home");
+            }
+            var context = new Team6Context();
+            string sqlQuery = "SELECT * FROM dbo.Criminal";
+            bool cont = false;
+
+            if (string.IsNullOrEmpty(firstName) && string.IsNullOrEmpty(lastName) && string.IsNullOrEmpty(weight) && string.IsNullOrEmpty(height) && string.IsNullOrEmpty(ssn) && string.IsNullOrEmpty(address) && string.IsNullOrEmpty(zipCode) && race.Equals(Race.None) && state.Equals(State.None) && gender.Equals(Gender.None) && eyeColor.Equals(EyeColor.None) && hairColor.Equals(HairColor.None))
+            {
+                // some error message that the user should enter at least one field
+            }
+            else
+            {
+                sqlQuery = "SELECT * FROM dbo.Criminal WHERE ";
+
+                if (!string.IsNullOrEmpty(firstName))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    firstName = firstName.Replace("'", "''");
+                    sqlQuery += "FirstName LIKE \'%" + firstName + "%\'";
+                    cont = true;
+                }
+
+                if (!string.IsNullOrEmpty(lastName))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    lastName = lastName.Replace("'", "''");
+                    sqlQuery += "LastName LIKE \'%" + lastName + "%\'";
+                    cont = true;
+                }
+
+                if (!string.IsNullOrEmpty(weight))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "Weight LIKE \'%" + Convert.ToInt32(weight) + "%\'";
+                    cont = true;
+                }
+
+                if (!string.IsNullOrEmpty(height))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "Height LIKE \'%" + Convert.ToInt32(height) + "%\'";
+                    cont = true;
+                }
+
+                if (!string.IsNullOrEmpty(ssn))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "Ssn LIKE \'%" + Convert.ToInt32(ssn) + "%\'";
+                    cont = true;
+                }
+
+                if (!string.IsNullOrEmpty(address))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    address = address.Replace("'", "''");
+                    sqlQuery += "Address LIKE \'%" + address + "%\'";
+                    cont = true;
+                }
+
+                if (!string.IsNullOrEmpty(zipCode))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "ZipCode LIKE \'%" + Convert.ToInt32(zipCode) + "%\'";
+                    cont = true;
+                }
+
+                if (!race.Equals(Race.None))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "Race = " + Convert.ToInt32(race);
+                    cont = true;
+                }
+
+                if (!state.Equals(State.None))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "State = " + Convert.ToInt32(state);
+                    cont = true;
+                }
+
+                if (!gender.Equals(Gender.None))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "Gender = " + Convert.ToInt32(gender);
+                    cont = true;
+                }
+
+                if (!eyeColor.Equals(EyeColor.None))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "EyeColor = " + Convert.ToInt32(eyeColor);
+                    cont = true;
+                }
+
+                if (!hairColor.Equals(HairColor.None))
+                {
+                    if (cont)
+                        sqlQuery += " AND ";
+                    sqlQuery += "HairColor = " + Convert.ToInt32(hairColor);
+                    cont = true;
+                }
+
+            }
+            
+            //using (StreamWriter _writer = new StreamWriter(fs))
+            //{
+            //    _writer.WriteLine(sqlQuery);
+            //}
+
+            var criminals = context.Criminals.SqlQuery(sqlQuery);
+            return View(criminals.OrderBy(x => x.FirstName).ThenBy(x => x.LastName).ToList());
         }
     }
 }
